pyproject.toml
[project]
name = "agent-support"
dynamic = ["version"]
description = "Supports agent operations by providing necessary tools and services."
readme = "README.md"
authors = [
    { name = "Sangram Reddy", email = "reddy.horcrux@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "async-lru==2.0.5",
    "fastapi==0.118.0",
    "firebase-admin==6.9.0",
    "google-adk==1.15.1",
    "google-cloud-firestore==2.21.0",
    "google-cloud-pubsub==2.31.1",
    "python-dotenv==1.1.1",
    "tenacity==8.5.0",
    "uvicorn==0.37.0",
    "yamale==6.0.0",
]

[project.scripts]
agent-support-service = "agent_support:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=24.10.0",
    "build>=1.2.2.post1",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
    "setuptools>=75.6.0",
    "jupyter>=1.1.1",
    "pyarrow>=19.0.1",
]

[tool.uv]
environments = [
    "sys_platform == 'darwin'",
    "sys_platform == 'linux'",
]

[tool.hatch.version]
path = "src/agent_support/__init__.py"
